libsmm_acc: time init and finalize functions
