+++ src/acc/libsmm_acc/libsmm_acc.h
@@ -24,12 +24,6 @@
 #include <unordered_map>
 #include <vector>
 
-extern "C" void dbcsr_timeset(const char** routineN, int* routineN_len, int* handle);
-void timeset(std::string routine_name, int& handle);
-
-extern "C" void dbcsr_timestop(int* handle);
-void timestop(int handle);
-
 enum libsmm_acc_algo {
     largeDB1 = 1,
     largeDB2 = 2,
