+++ src/acc/libsmm_acc/libsmm_acc_init.h
@@ -13,6 +13,13 @@
 #include "../cuda/acc_blas.h"
 
 #include <vector>
+#include <string>
+
+extern "C" void dbcsr_timeset(const char** routineN, int* routineN_len, int* handle);
+void timeset(std::string routine_name, int& handle);
+
+extern "C" void dbcsr_timestop(int* handle);
+void timestop(int handle);
 
 extern "C" int libsmm_acc_init (void);
